# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: OriginLock API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List user projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /verify/{certificateId}:
    get:
      operationId: verifyCertificate
      tags: [verification]
      summary: Get certificate verification data
      parameters:
        - name: certificateId
          in: path
          required: true
          schema:
            type: string
        - name: submittedHash
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResponse"
        "404":
          description: Certificate not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    ProjectResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        fileCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - fileCount
        - createdAt
        - updatedAt
    ProjectDetailResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        fileCount:
          type: integer
        files:
          type: array
          items:
            $ref: "#/components/schemas/FileRecordResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - fileCount
        - files
        - createdAt
        - updatedAt
    CreateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
      required:
        - title
    UpdateProjectBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
    FileRecordResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        projectId:
          type: ["string", "null"]
        projectTitle:
          type: ["string", "null"]
        title:
          type: string
        originalFilename:
          type: string
        fileType:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: string
        sha256Hash:
          type: string
        notes:
          type: ["string", "null"]
        privacyMode:
          type: string
          enum: [private, shareable_certificate, public_verification]
        certificateId:
          type: ["string", "null"]
        certificateStatus:
          type: ["string", "null"]
        verificationUrl:
          type: ["string", "null"]
        recordedAtUtc:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - originalFilename
        - fileType
        - mimeType
        - fileSizeBytes
        - sha256Hash
        - privacyMode
        - createdAt
        - updatedAt
...
    VerificationResponse:
      type: object
      properties:
        certificateId:
          type: string
        status:
          type: string
          enum: [valid, revoked, superseded]
        displayName:
          type: string
        creatorHandle:
          type: string
        projectTitle:
          type: ["string", "null"]
        fileTitle:
          type: string
        originalFilename:
          type: string
        fileType:
          type: string
        mimeType:
          type: string
        fileSizeBytes:
          type: string
        sha256Hash:
          type: string
        recordedAtUtc:
          type: string
          format: date-time
        hashMatchResult:
          type: ["string", "null"]
          enum: [match, mismatch, null]
      required:
        - certificateId
        - status
        - displayName
        - creatorHandle
        - fileTitle
        - originalFilename
        - fileType
        - mimeType
        - fileSizeBytes
        - sha256Hash
        - recordedAtUtc
...