# lib/api-zod/src/generated/types/updateProjectBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OriginLock API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateProjectBody {
  title?: string;
  /** @nullable */
  description?: string | null;
}