# lib/api-zod/src/generated/types/verificationResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OriginLock API specification
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationResponseHashMatchResult } from "./verificationResponseHashMatchResult";
import type { VerificationResponseStatus } from "./verificationResponseStatus";

export interface VerificationResponse {
  certificateId: string;
  status: VerificationResponseStatus;
  displayName: string;
  creatorHandle: string;
  /** @nullable */
  projectTitle?: string | null;
  fileTitle: string;
  originalFilename: string;
  fileType: string;
  mimeType: string;
  fileSizeBytes: string;
  sha256Hash: string;
  recordedAtUtc: Date;
  /** @nullable */
  hashMatchResult?: VerificationResponseHashMatchResult;
}